Move config into it's own file
Mostly a copy and paste of config functionality into it's own file.
Cleans up the package a little.

Minor change:

* Allow the Kinesis Client to be initialized in defaults if not set
