+++ kinesis.go
@@ -7,72 +7,18 @@ import (
 
 	"github.com/apex/log"
 	k "github.com/aws/aws-sdk-go/service/kinesis"
-	"github.com/aws/aws-sdk-go/service/kinesis/kinesisiface"
-	"github.com/jpillora/backoff"
 )
 
 // Size limits as defined by http://docs.aws.amazon.com/kinesis/latest/APIReference/API_PutRecords.html.
 const (
-	maxRecordSize        = 1 << 20 // 1MiB
-	maxRequestSize       = 5 << 20 // 5MiB
-	maxRecordsPerRequest = 500
+	maxRecordSize  = 1 << 20 // 1MiB
+	maxRequestSize = 5 << 20 // 5MiB
 )
 
 // Errors.
 var (
 	ErrRecordSizeExceeded = errors.New("kinesis: record size exceeded")
 )
-
-type Config struct {
-	// StreamName is the Kinesis stream.
-	StreamName string
-
-	// FlushInterval is a regular interval for flushing the buffer. Defaults to 1s.
-	FlushInterval time.Duration
-
-	// BufferSize determines the batch request size. Must not exceed 500. Defaults to 500.
-	BufferSize int
-
-	// BacklogSize determines the channel capacity before Put() will begin blocking. Defaults to 500.
-	BacklogSize int
-
-	// Backoff determines the backoff strategy for record failures.
-	Backoff backoff.Backoff
-
-	// Logger is the logger used. Defaults to log.Log.
-	Logger log.Interface
-
-	// Client is the Kinesis API implementation.
-	Client kinesisiface.KinesisAPI
-}
-
-// defaults for configuration.
-func (c *Config) defaults() {
-	if c.Logger == nil {
-		c.Logger = log.Log
-	}
-
-	c.Logger = c.Logger.WithFields(log.Fields{
-		"package": "kinesis",
-		"stream":  c.StreamName,
-	})
-
-	if c.BufferSize == 0 {
-		c.BufferSize = maxRecordsPerRequest
-	}
-
-	if c.BufferSize > maxRecordsPerRequest {
-		panic("kinesis: BufferSize exceeds 500")
-	}
-
-	if c.BacklogSize == 0 {
-		c.BacklogSize = maxRecordsPerRequest
-	}
-
-	if c.FlushInterval == 0 {
-		c.FlushInterval = time.Second
-	}
-}
 
 // Producer batches records.
 type Producer struct {
